requirements.txt
django>=1.7.0
Pillow>=2.7.0
django-crispy-forms